LayoutTests/inspector/console/x-frame-options-message-expected.txt

CONSOLE MESSAGE: The X-Frame-Option 'deny' supplied in a <meta> element was ignored. X-Frame-Options may only be provided by an HTTP header sent with the document.


== Running test suite: Console.XFrameOptionsMessages
-- Running test case: XFrameOptionsDeny
PASS: The message should be due to a security violation.
PASS: The message should be an error.
PASS: The message should come from the main target.
